backend/requirements.txt
fastapi==0.86.0
pydantic==1.10.2
uvicorn==0.19.0
sqlmodel==0.0.8
#sqlalchemy==1.4.43
psycopg2-binary==2.9.5
python-dotenv==0.21.0